feat: Update nutrition table to use decimal for calories; rename profile_user table to profile_user_pivot
feat: Enhance experiment component with new navigation and dropdown features

feat: Add footer component with brand, links, and social media icons

feat: Revamp navbar with mobile support and user authentication options

feat: Implement new page routes and create PageController for static pages

feat: Create About, Community, Contact Us, Culinary Resources, and Educational Resources views

fix: Update test view with new navigation structure

+++ database/migrations/2025_09_06_151822_create_nutrition_table.php
@@ -12,7 +12,7 @@ return new class extends Migration {
     {
         Schema::create('nutrition', function (Blueprint $table) {
             $table->foreignId('recipe_id')->primary()->constrained('recipes')->cascadeOnDelete();
-            $table->integer('calories')->nullable();
+            $table->decimal('calories', 6, 2)->nullable();
             $table->decimal('fat', 6, 2)->nullable();
             $table->decimal('carbs', 6, 2)->nullable();
             $table->decimal('protein', 6, 2)->nullable();
